+++ docs/CHANGELOG.md
@@ -80,6 +80,7 @@ include_toc: true
     * `Improve API Chat Messages` (Enabled By Default)
     * `Sync Chunks To Disk After Saving` (Enabled By Default)
     * `Implement "Quit and copy map" Button` (Enabled By Default)
+    * `Only Send Messages To Fully Connected Players` (Enabled By Default)
   * New Bug Fixes (All Enabled By Default)
     * `Fix Held Item Caching`
     * `Fix Pigmen Burning In The Sun`